# src/components/setOfQuestions/setOfQuestions.types.ts
export interface ISetOfQuestions {
    name: string;
    description: string;
    criteriaId: string;
}

export interface ISetOfQuestionsDTO {
    id: string,
    name: string,
    description: string,
    criteriaId: string,
}