# swagger.json
{
	"components": {
...
		"schemas": {
			"IUserRegister": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"confirmPassword": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"facilityId": {
						"type": "string"
					},
					"positionId": {
						"type": "string"
					}
				},
				"required": [
					"email",
					"password",
					"confirmPassword",
					"name",
					"phone",
					"address",
					"facilityId",
					"positionId"
				],
				"type": "object",
				"additionalProperties": true
			},
			"IUserLogin": {
				"properties": {
					"email": {
						"type": "string"
					},
					"password": {
						"type": "string"
					},
					"remember": {
						"type": "boolean"
					}
				},
				"required": [
					"email",
					"password",
					"remember"
				],
				"type": "object",
				"additionalProperties": true
			},
			"IRefreshTokenReq": {
				"properties": {
					"refreshToken": {
						"type": "string"
					}
				},
				"required": [
					"refreshToken"
				],
				"type": "object",
				"additionalProperties": true
			},
			"IUserProfile": {
				"properties": {
					"name": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"facilityId": {
						"type": "string"
					},
					"positionId": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"name",
					"phone",
					"address",
					"facilityId",
					"positionId",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": true
			},
			"IAccessTokenReq": {
				"properties": {
					"accessToken": {
						"type": "string"
					}
				},
				"required": [
					"accessToken"
				],
				"type": "object",
				"additionalProperties": true
			},
			"IUserUpdateProfile": {
				"properties": {
					"name": {
						"type": "string"
					},
					"phone": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"facilityId": {
						"type": "string"
					},
					"positionId": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"phone",
					"address",
					"facilityId",
					"positionId"
				],
				"type": "object",
				"additionalProperties": true
			},
			"IUserUpdatePassword": {
				"properties": {
					"oldPassword": {
						"type": "string"
					},
					"newPassword": {
						"type": "string"
					},
					"confirmPassword": {
						"type": "string"
					}
				},
				"required": [
					"oldPassword",
					"newPassword",
					"confirmPassword"
				],
				"type": "object",
				"additionalProperties": true
			},
...
			"ISetOfQuestions": {
				"properties": {
					"name": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"criteriaId": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"description",
					"criteriaId"
				],
				"type": "object",
				"additionalProperties": true
			},
			"IQuestion": {
				"properties": {
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"content": {
						"type": "string"
					},
					"setQuestionId": {
						"type": "string"
					}
				},
				"required": [
					"title",
					"description",
					"content",
					"setQuestionId"
				],
				"type": "object",
				"additionalProperties": true
			},
...
		},
		"securitySchemes": {
			"api_key": {
				"type": "apiKey",
				"name": "access_token",
				"in": "query"
			},
			"jwt": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	},
	"info": {
		"title": "uinnovate",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/users/register": {
			"post": {
				"operationId": "Register",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUserRegister"
							}
						}
					}
				}
			}
		},
		"/users/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUserLogin"
							}
						}
					}
				}
			}
		},
		"/users/refreshtoken": {
			"post": {
				"operationId": "RefreshToken",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IRefreshTokenReq"
							}
						}
					}
				}
			}
		},
		"/users/getuserinfo": {
			"post": {
				"operationId": "GetUserInfo",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/IUserProfile"
										},
										{}
									]
								}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IAccessTokenReq"
							}
						}
					}
				}
			}
		},
		"/users/profile": {
			"get": {
				"operationId": "GetProfile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			},
			"put": {
				"operationId": "UpdateProfile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUserUpdateProfile"
							}
						}
					}
				}
			}
		},
		"/users/changepassword": {
			"put": {
				"operationId": "ChangePassword",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Users"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IUserUpdatePassword"
							}
						}
					}
				}
			}
		},
...
		"/setOfQuestions": {
			"post": {
				"operationId": "CreateSetOfQuestions",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"SetOfQuestions"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ISetOfQuestions"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetSetOfQuestions",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"SetOfQuestions"
				],
				"security": [],
				"parameters": []
			}
		},
		"/setOfQuestions/{id}": {
			"put": {
				"operationId": "UpdateSetOfQuestions",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"SetOfQuestions"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/ISetOfQuestions"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteSetOfQuestions",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"SetOfQuestions"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/questions/{setQuestionId}": {
			"get": {
				"operationId": "GetQuestionBySetQuestionId",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"description": "Get question by setQuestionId",
				"tags": [
					"Questions"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "setQuestionId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/questions/{id}": {
			"put": {
				"operationId": "UpdateQuestion",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Questions"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IQuestion"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteQuestion",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Questions"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}